requirements.txt
biopython==1.72
importlib-resources==5.2.2
ipywidgets==7.6.3
jupyterlab-widgets==1.0.0
Keras==2.3.1
scipy==1.5.4
smmap==4.0.0
streamlit==1.0.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
